interview round two tweaks (#57)
## Summary

* recently-added interview questions **not** included in interview review form after round one

* interview question fields allow `NULL` to differentiate between whether interviewer was presented with (question) field or not

## Change sets

* added sqlmigrate to list of supported db commands proxied into local container

* made InterviewReview question fields NULL-y

(in hopes of enabling differentiation between when user presented with
questions -- i.e. in round one -- but perhaps didn't enter anything, vs.
when user not presented with questions -- i.e. in subsequent rounds.)

* made interview review form include interview questions ONLY during round one

(and tweaked presentation of field groups, to omit the remaining questions' fieldset legend, since that's a bit much.)